Progress: Revisionn CRUD Humas Page

+++ app/Http/Controllers/HumasController.php
@@ -30,8 +30,10 @@ class HumasController extends Controller
         ]);
 
         if ($request->file('banner')) {
-            $oldImagePath = public_path('assets/img/humas-images/header-image/') . $request->oldImage;
-            unlink($oldImagePath);
+            if (public_path('assets/img/humas-images/header-image/') . $request->oldImage && $request->oldImage) {
+                $oldImagePath = public_path('assets/img/humas-images/header-image/') . $request->oldImage;
+                unlink($oldImagePath);
+            }
 
             $image = $request->file('banner');
             $imageName = time() . '.' . $image->getClientOriginalExtension();
